modulized record python scripts
